requirements_dev.txt
requests==2.22.0
pandas==0.25.3
appdirs==1.4.3
setuptools==41.2.0
packaging==19.0
Click==7.0
xmltodict==0.12.0
pytest==4.6.6
zeep==3.4.0
ddt==1.2.1
diskcache==4.1.0
numpy==1.17.4
python_dateutil==2.8.1
teamcity_messages==1.27
typing==3.7.4.1
